# openapi.yaml
openapi: 3.0.3
info:
  title: API сервиса для тестового задания
  version: 0.1.0
tags:
  - name: cats
    description: Всё что касается котиков
paths:
  /cats/like/{uid}:
    post:
      tags:
        - cats
      summary: Добавление лайка
      operationId: addLike
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
          description: Уникальный идентификатор картинки
      requestBody:
        description: URL котика
        content:
          application/json:
            schema:
              type: object
              required:
                - url
              properties:
                url:
                  type: string
                  description: URL котика
        required: true
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cat'
        '405':
          description: Invalid input
    delete:
      tags:
        - cats
      summary: Удаление лайка
      operationId: removeLike
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
          description: Уникальный идентификатор картинки
      responses:
        '200':
          description: Successful operation
        '404':
          description: Like not found
  /cats/likes:
    get:
      tags:
        - cats
      summary: Список лайков
      operationId: getAllLikes
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cat'
  /user:
    post:
      tags:
        - users
      summary: Добавление пользователя
      operationId: newUser
      requestBody:
        description: Новый пользователь
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '201':
          description: Successful operation
          headers:
            X-Auth-Token:
              description: Токен для последующей авторизации, например sha256(user_id + secret_salt)
              example: 3525dcdddea774939652f7f11df6d7db10a9db35a5d758c64d600a00c1cc41be
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '405':
          description: Invalid input
        '409':
          description: User with this login already exists
  /user/login:
    post:
      tags:
        - users
      summary: Авторизация пользователя
      operationId: loginUser
      requestBody:
        description: Данные для входа
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginUser'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: Токен для последующей авторизации
        '401':
          description: Invalid login or password

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Ожидается токен, выдаваемый при регистрации пользователя
      bearerFormat: SHA256
  schemas:
    User:
      type: object
      required:
        - login
        - password
      properties:
        login:
          type: string
          description: Логин пользователя
        password:
          type: string
          description: Пароль пользователя
    LoginUser:
      type: object
      required:
        - login
        - password
      properties:
        login:
          type: string
          description: Логин пользователя
        password:
          type: string
          description: Пароль пользователя